# Use UnboundMethod#bind_call method instead to improve performance